+++ src/ui/map_settings/panel.py
@@ -19,6 +19,7 @@ from ...qt_compat import (
     SIZE_POLICY_PREFERRED,
     SPIN_BOX_UP_DOWN_ARROWS,
     SCROLL_BAR_ALWAYS_OFF,
+    configure_compact_command_button,
     available_screen_geometry,
     screen_aware_popup_position,
 )
@@ -36,7 +37,6 @@ class MapSettingsPanel(QtWidgets.QWidget):
     CONTENT_MAXIMUM_WIDTH = 280
     BUTTON_SIZE = 24
     ICON_SIZE = 20
-    HOVER_STYLE = "QPushButton:hover {\n    border: 1px solid #bbb;\n}\n"
     TOGGLE_STYLE = """
 QPushButton {
     border: 1px solid transparent;
@@ -112,6 +112,9 @@ QPushButton {
             self.cb_symbol_continuous_colormap,
             self.sb_symbol_classes,
             self.sb_symbol_size,
+            self.cmb_symbol_marker_shape,
+            self.pb_symbol_outline_color,
+            self.sb_symbol_outline_width,
             self.sb_symbol_opacity,
             self.cb_symbology_live,
             self.pb_symbology,
@@ -167,6 +170,15 @@ QPushButton {
         self.cb_symbol_continuous_colormap = (
             self.symbology_settings_popup.cb_symbol_continuous_colormap
         )
+        self.cmb_symbol_marker_shape = (
+            self.symbology_settings_popup.cmb_symbol_marker_shape
+        )
+        self.pb_symbol_outline_color = (
+            self.symbology_settings_popup.pb_symbol_outline_color
+        )
+        self.sb_symbol_outline_width = (
+            self.symbology_settings_popup.sb_symbol_outline_width
+        )
 
         self.cb_symbology_live = QtWidgets.QCheckBox(self)
         self.cb_symbology_live.setObjectName("cb_symbology_live")
@@ -209,7 +221,7 @@ QPushButton {
             value=0.0,
             single_step=0.1,
         )
-        self._configure_flat_button(
+        self._configure_command_button(
             self.pb_reference_reset,
             icon=":/icons/icons/reset_offset.svg",
             tooltip="Reset Reference to zero",
@@ -225,7 +237,7 @@ QPushButton {
             self._sync_reference_editability
         )
         self._sync_reference_editability()
-        self._configure_flat_button(
+        self._configure_command_button(
             self.pb_symbol_range_settings,
             icon=":/icons/icons/edit.svg",
             tooltip="Configure range source and symmetry",
@@ -250,7 +262,7 @@ QPushButton {
             tooltip="Reverse colormap",
             checked=False,
         )
-        self._configure_flat_button(
+        self._configure_command_button(
             self.pb_symbology_settings,
             icon=":/icons/icons/setting.svg",
             tooltip="Configure symbology",
@@ -517,16 +529,20 @@ QPushButton {
         spin_box.setSizePolicy(horizontal_policy, SIZE_POLICY_FIXED)
 
     def _configure_toggle_button(self, button, *, icon, tooltip, checked):
-        self._configure_button_base(button, icon=icon, tooltip=tooltip)
-        button.setFlat(True)
+        button.setText("")
+        button.setIcon(QtGui.QIcon(icon))
+        button.setToolTip(tooltip)
+        button.setAccessibleName(tooltip)
         button.setCheckable(True)
         button.setChecked(checked)
+        button.setFlat(True)
+        button.setFixedSize(self.BUTTON_SIZE, self.BUTTON_SIZE)
+        button.setIconSize(QSize(self.ICON_SIZE, self.ICON_SIZE))
+        button.setSizePolicy(SIZE_POLICY_FIXED, SIZE_POLICY_FIXED)
         button.setStyleSheet(self.TOGGLE_STYLE)
 
-    def _configure_flat_button(self, button, *, icon, tooltip):
+    def _configure_command_button(self, button, *, icon, tooltip):
         self._configure_button_base(button, icon=icon, tooltip=tooltip)
-        button.setFlat(True)
-        button.setStyleSheet(self.HOVER_STYLE)
 
     def _show_range_settings_popup(self, checked=False):
         """Open the compact range settings popup next to its anchor button."""
@@ -610,6 +626,8 @@ QPushButton {
         button.setIcon(QtGui.QIcon(icon))
         button.setToolTip(tooltip)
         button.setAccessibleName(tooltip)
-        button.setMaximumSize(self.BUTTON_SIZE, self.BUTTON_SIZE)
-        button.setIconSize(QSize(self.ICON_SIZE, self.ICON_SIZE))
-        button.setSizePolicy(SIZE_POLICY_FIXED, SIZE_POLICY_FIXED)
+        configure_compact_command_button(
+            button,
+            size=self.BUTTON_SIZE,
+            icon_size=self.ICON_SIZE,
+        )